# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Deckwatch API
  description: |
    Kubernetes control-plane and OCI registry surface exposed by Deckwatch.

    - `/api/*` — application, deployment, namespace, and cluster management
    - `/v2/*`  — OCI Distribution Spec v1.1 (only when the embedded registry is enabled)
    - `/metrics` — Prometheus exposition
  version: "0.1.0"
  contact:
    name: Deckwatch
    url: https://github.com/kore/deckwatch
servers:
  - url: /
    description: Same-origin (served by deckwatch)
...
paths:
...
  /api/namespaces/{ns}/deployments/{name}/addons/{addon_id}:
    parameters:
      - $ref: "#/components/parameters/Namespace"
      - $ref: "#/components/parameters/Name"
      - in: path
        name: addon_id
        required: true
        schema: { type: string }
    post:
      tags: [Addons]
      summary: Attach an addon to a deployment
      responses: { "200": { description: OK } }
    patch:
      tags: [Addons]
      summary: Update attached addon config
      responses: { "200": { description: OK } }
    delete:
      tags: [Addons]
      summary: Detach an addon
      responses: { "204": { description: Detached } }
...
components:
  parameters:
    Namespace:
      in: path
      name: ns
      required: true
      schema: { type: string }
      description: Kubernetes namespace
    Name:
      in: path
      name: name
      required: true
      schema: { type: string }
      description: Resource name
...